refactor(cisco): isolate scan containment
