+++ src/codex_plugin_scanner/guard/runtime/shell_execution_context.py
@@ -418,6 +418,7 @@ def _identity_payload(identity: ShellPathIdentity | None) -> dict[str, int] | No
         return None
     return {
         "change_time_ns": identity.change_time_ns,
+        "creation_time_ns": identity.creation_time_ns,
         "device": identity.device,
         "inode": identity.inode,
         "mode": identity.mode,
